Move empty state to TableEmptyState component
