+++ packages/design-system/src/components/Table/TableEmptyState/TableEmptyState.tsx
@@ -5,23 +5,20 @@
 //
 // SPDX-License-Identifier: MIT
 //
-import type { DataTableProps } from './DataTable'
-import { EmptyState as EmptyStateBase } from '../EmptyState'
+import { EmptyState, type EmptyStateProps } from '../../EmptyState'
 import { TableCell, TableRow } from '../Table'
 
-interface EmptyStateProps extends Pick<DataTableProps<unknown>, 'entityName'> {
-  globalFilter?: string
+interface TableEmptyStateProps extends EmptyStateProps {
   colSpan: number
 }
 
-export const EmptyState = ({
-  entityName,
+export const TableEmptyState = ({
   colSpan,
-  globalFilter,
-}: EmptyStateProps) => (
+  ...props
+}: TableEmptyStateProps) => (
   <TableRow isHoverable={false}>
     <TableCell colSpan={colSpan} className="h-24 text-center">
-      <EmptyStateBase textFilter={globalFilter} entityName={entityName} />
+      <EmptyState {...props} />
     </TableCell>
   </TableRow>
 )
