+++ packages/design-system/src/components/Table/TableEmptyState/index.tsx
@@ -1,0 +1,8 @@
+//
+// This source file is part of the Stanford Biodesign Digital Health ENGAGE-HF open-source project
+//
+// SPDX-FileCopyrightText: 2023 Stanford University and the project authors (see CONTRIBUTORS.md)
+//
+// SPDX-License-Identifier: MIT
+//
+export * from './TableEmptyState'
